+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - merge default list of excluded attributes with custom configuration
 - add `autocomplete` to the list of excluded attributes
-- support replace text selection on `enter` 
+- support replace text selection on `enter`
+- support replace text selection on `paste`
 
 # 0.0.27 (2022-06-05)